Fix pop-up dispaly
